+++ packages/analyzer/src/test/static/script-tags/modules/usesClass.js
@@ -1,0 +1,4 @@
+// This should not work
+class Bar extends Foo {
+
+}