+++ packages/analyzer/CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Dropped support for node v6. This is a soft break, as we aren't
   making any changes that are known to break node v6, but we're no longer testing against it. See our [node version support policy](https://www.polymer-project.org/2.0/docs/tools/node-support)
   for details.
+* ScriptTagImport#isModule is true for `<script type="module" src="...">`
+* Module `<script>` imports don't inherit imports from the HTML file that
+  imports them, because unlike normal scripts, modules can be imported from
+  many HTML files, and they have their own way to import their dependencies,
+  rather than depending on HTML container files.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.24] - 2018-04-18